feat: accessToken을 zustand로 가져오도록 수정
