+++ src/modules/fetchClient.ts
@@ -1,4 +1,5 @@
 import { processEnvBaseUrl } from 'constants/environments';
+import { initializeStore } from 'stores/createAppStore';
 
 export interface ApiRequest {
   url: string;
@@ -71,10 +72,8 @@ const request = async <TResponse>(
   { url, params, body }: ApiRequest,
   init: Omit<RequestInit, 'credentials'> = {}
 ) => {
-  // TODO: vanilla/zustand를 통해 accessToken 가져오기
-  // const accessToken = sessionStore.getState().getAccessToken();
-  const token =
-    'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOjEsInByb3ZpZGVyIjoiS0FLQU8iLCJ0eXBlIjoiQUNDRVNTIiwiaWF0IjoxNzQ5OTY5MzI0LCJleHAiOjE3NTc3NDUzMjR9.bDpurCfyQ906gPYbPzEnOkzoZpBxLElwXjKY3rwWj9Q';
+  const sessionStore = initializeStore();
+  const token = sessionStore.getState().accessToken;
 
   const headers = new Headers(init.headers);
   if (token) headers.set('Authorization', `Bearer ${token}`);
